LayoutTests/js/Promise-resolve-with-then-fulfill-expected.txt

Test Promise.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS value.then is called.
PASS thisValue is value
The promise is not fulfilled now.
PASS fulfilled
PASS result is "hello"
PASS successfullyParsed is true

TEST COMPLETE
